pyproject.toml
[project]
name = "agentwatch"
version = "0.1.0"
description = "Few-shot silent failure detection for AI agents — materialized views for natural language over production traces"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # ML core
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "sentence-transformers>=3.0.0",
    "setfit>=1.0.0",
    "datasets>=2.19.0",
    "scikit-learn>=1.4.0",
    "hdbscan>=0.8.33",
    "faiss-cpu>=1.8.0",
    # Agent
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    # Utilities
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
]

[project.optional-dependencies]
gpu = ["faiss-gpu>=1.8.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
line-length = 100
target-version = "py310"

